Add login status to cache and update initial route
The commit adds a new boolean value to the cache to store the login status of the user. The `LoginCubit` has been updated to set this value when the user logs in or fails to log in. Additionally, the `startScreen` method has been updated to return the appropriate initial route based on the login status stored in the cache. Finally, the `initialRoute` in the `AdminRebuyApp` has been updated to use the